KN-#699bI3vZ: setup  project
- install require packages and libraries
- intagrate code qality checker
- orginize folder structure

delivers(chore): #699bI3vZ
