+++ .circleci/config.yml
@@ -7,7 +7,7 @@ jobs:
   build:
     docker:
       # specify the version you desire here
-      - image: circleci/node:7.10
+      - image: circleci/node:12.14.1
 
       # Specify service dependencies here if necessary
       # CircleCI maintains a library of pre-built images
