+++ .codeclimate.yml
@@ -1,0 +1,43 @@
+version: '2'
+checks:
+  argument-count:
+    config:
+      threshold: 4
+  complex-logic:
+    config:
+      threshold: 4
+  file-lines:
+    config:
+      threshold: 250
+  method-complexity:
+    config:
+      threshold: 5
+  method-count:
+    config:
+      threshold: 50
+  method-lines:
+    config:
+      threshold: 25
+  nested-control-flow:
+    config:
+      threshold: 4
+  return-statements:
+    config:
+      threshold: 4
+  similar-code:
+    config:
+      threshold: # language-specific defaults. an override will affect all languages.
+  identical-code:
+    config:
+      threshold: # language-specific defaults. an override will affect all languages.
+plugins:
+  eslint:
+    enabled: true
+exclude_patterns:
+  - '**/node_modules/'
+  - 'script/'
+  - '**/spec/'
+  - '**/tests/'
+  - '**/__tests__/'
+  - '**/vendor/'
+  - '**/*.d.ts'
