+++ .gitignore
@@ -117,3 +117,4 @@ dist
 
 # TernJS port file
 .tern-port
+.vscode
