+++ README.md
@@ -1,4 +1,4 @@
-[![CircleCI](https://circleci.com/gh/kinyarwanda-net/python-backend/tree/develop.svg?style=svg)](https://circleci.com/gh/kinyarwanda-net/python-backend/tree/develop)
+[![CircleCI](https://circleci.com/gh/kinyarwanda-net/mobile-app/tree/develop.svg?style=svg)](https://circleci.com/gh/kinyarwanda-net/mobile-app/tree/develop)
 
 mobile-app
 React native mobile app
