+++ babel.config.js
@@ -1,6 +1,19 @@
-module.exports = function(api) {
+module.exports = api => {
   api.cache(true);
   return {
     presets: ['babel-preset-expo'],
+    plugins: [
+      [
+        'module-resolver',
+        {
+          root: ['./src'],
+          alias: {
+            src: './src',
+            test: './tests',
+            underscore: 'lodash',
+          },
+        },
+      ],
+    ],
   };
 };
