--- components/StyledText.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import { Text } from 'react-native';
-
-export function MonoText(props) {
-  return (
-    <Text {...props} style={[props.style, { fontFamily: 'space-mono' }]} />
-  );
-}
