+++ config.js
@@ -1,0 +1,3 @@
+import { CLIENT_ID } from 'react-native-dotenv';
+
+export default { CLIENT_ID };
