+++ jest.config.js
@@ -1,0 +1,31 @@
+const { defaults } = require('jest-config');
+
+module.exports = {
+  displayName: 'kinyarwanda',
+  verbose: true,
+  preset: 'jest-expo',
+  testMatch: ['**/?(*.)+(spec|test).js'],
+  collectCoverage: true,
+  clearMocks: true,
+  collectCoverageFrom: [
+    'src/**/*.{js,jsx}',
+    '!**/coverage/**',
+    '!**/node_modules/**',
+    '!**/babel.config.js',
+    '!**/jest.setup.js',
+    '!src/constants/**',
+    '!src/navigation/**',
+  ],
+
+  setupFiles: ['<rootDir>/jest/setup.js'],
+  snapshotSerializers: ['enzyme-to-json/serializer'],
+  moduleFileExtensions: [
+    ...defaults.moduleFileExtensions,
+    'js',
+    'json',
+    'jsx',
+    'ts',
+    'tsx',
+    'node',
+  ],
+};
