+++ jsconfig.json
@@ -1,0 +1,6 @@
+{
+  "compilerOptions": {
+    "module": "es6",
+    "baseUrl": "./src"
+  }
+}
