+++ src/constants/Layout.js
@@ -1,7 +1,7 @@
 import { Dimensions } from 'react-native';
 
-const width = Dimensions.get('window').width;
-const height = Dimensions.get('window').height;
+const { width } = Dimensions.get('window');
+const { height } = Dimensions.get('window');
 
 export default {
   window: {
