+++ src/navigation/AppNavigator.js
@@ -1,0 +1,16 @@
+import { createStackNavigator } from 'react-navigation-stack';
+import HomeScreen from '../screens/HomeScreen';
+
+const navigator = createStackNavigator(
+  {
+    Home: HomeScreen,
+  },
+  {
+    initialRouteName: 'Home',
+    defaultNavigationOptions: {
+      title: 'Kinyarwanda',
+    },
+  },
+);
+
+export default navigator;
