+++ src/screens/HomeScreen.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { Text, View } from 'react-native';
+import styles from 'src/styles/homeScreenStyle';
+
+const HomeScreen = () => (
+  <View style={styles.container}>
+    <Text>Hello World</Text>
+  </View>
+);
+
+export default HomeScreen;
